touristapp_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
idna==3.10
motor==3.7.1
passlib[bcrypt]==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
aiohttp==3.9.1
aiofiles==23.2.1
Pillow>=10.0.0
setuptools>=68.0.0
wheel>=0.40.0
